Moved 2D Scan Analysis to `common` experiment
