+++ ScanAnalysis/scan_analysis/mapping/map_Undulator.py
@@ -5,7 +5,7 @@ from scan_analysis.analyzers.Undulator.mag_spec_stitcher_analysis import MagSpec
 from scan_analysis.analyzers.Undulator.rad2_spec_analysis import Rad2SpecAnalysis
 from scan_analysis.analyzers.Undulator.visa_ebeam_analysis import VisaEBeamAnalysis
 from scan_analysis.analyzers.Undulator.camera_image_analysis import CameraImageAnalysis
-from scan_analysis.analyzers.Undulator.array2D_scan_analysis import Array2DScanAnalysis
+from scan_analysis.analyzers.common.array2D_scan_analysis import Array2DScanAnalysis
 from scan_analysis.analyzers.Undulator.hamaspectro_analysis import FiberSpectrometerAnalysis
 from scan_analysis.analyzers.Undulator.frog_analysis import FrogAnalysis
